refactor: into selector
